server/requirements.txt
openenv-core[core]>=0.2.0
fastapi>=0.115.0
uvicorn>=0.24.0
numpy>=1.24.0
scipy>=1.10.0
pydantic>=2.0.0
